refactor(api): rename files with same convention
