+++ api/db/seeds/users.js
@@ -1,4 +1,4 @@
-import { createPassword } from "../../src/authentication/services/password.service.js";
+import { createPassword } from "../../src/authentication/services/password-service.js";
 import { USER_TYPES } from "../../src/shared/constants.js";
 import { databaseBuilder } from "../database-builder/index.js";
