+++ api/src/authentication/controllers/activate-user-controller.js
@@ -1,8 +1,8 @@
 import { ERRORS } from "../../shared/constants.js";
 import { logger } from "../../shared/logger.js";
 import { findUserById } from "../../shared/repositories/users-repository.js";
-import { activateUserByUserId } from "../repositories/register.repository.js";
-import { decodedToken } from "../services/token.service.js";
+import { activateUserByUserId } from "../repositories/register-repository.js";
+import { decodedToken } from "../services/token-service.js";
 
 async function activateUserController(req, res) {
   if (!req.headers.authorization) {
