+++ api/src/authentication/controllers/register-controller.js
@@ -2,10 +2,10 @@ import Joi from "joi";
 
 import { USER_TYPES } from "../../shared/constants.js";
 import { logger } from "../../shared/logger.js";
-import { checkSchema } from "../../shared/middlewares/checkSchema.js";
-import * as registerRepository from "../repositories/register.repository.js";
-import { createPassword } from "../services/password.service.js";
-import { sendMailToActivateUserService } from "../services/send-mail-to-activate-user.service.js";
+import { checkSchema } from "../../shared/middlewares/check-schema.js";
+import * as registerRepository from "../repositories/register-repository.js";
+import { createPassword } from "../services/password-service.js";
+import { sendMailToActivateUserService } from "../services/send-mail-to-activate-user-service.js";
 
 const createUserSchema = Joi.object({
   firstname: Joi.string().required(),
