+++ api/tests/authentication/integration/repositories/register-repository.test.js
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 
 import { databaseBuilder } from "../../../../db/database-builder/index.js";
 import { knex } from "../../../../db/knex-database-connection.js";
-import * as registerRepository from "../../../../src/authentication/repositories/register.repository.js";
+import * as registerRepository from "../../../../src/authentication/repositories/register-repository.js";
 import { USER_TYPES } from "../../../../src/shared/constants.js";
 
 describe("Integration | Authentication | Repository | Register", () => {
