--- api/tests/authentication/unit/services/send-mail-to-activate-user.service.spec.js
@@ -1,34 +1,0 @@
-import { describe, expect, it, vi } from "vitest";
-
-import {
-  sendMailToActivateUserService,
-} from "../../../../src/authentication/services/send-mail-to-activate-user.service.js";
-import { encodedToken } from "../../../../src/authentication/services/token.service.js";
-import { createMailBody } from "../../../../src/shared/services/emails/create-mail-body.js";
-import { sendMail } from "../../../../src/shared/services/emails/send-mail.js";
-
-vi.mock("../../../../src/authentication/services/token.service.js");
-vi.mock("../../../../src/shared/services/emails/create-mail-body.js");
-vi.mock("../../../../src/shared/services/emails/send-mail.js");
-
-describe("Unit | Authentication | Services | Send mail to activate user service", () => {
-  it("should create token, create mail body and send mail", async () => {
-    // given
-    const data = {
-      firstname: "John",
-      lastname: "Doe",
-      email: "john.doe@example.net",
-    };
-    const userId = 123;
-    encodedToken.mockResolvedValue("token");
-    createMailBody.mockResolvedValue("document");
-
-    // when
-    await sendMailToActivateUserService({ ...data, userId });
-
-    // then
-    expect(encodedToken).toHaveBeenCalledWith({ userId });
-    expect(createMailBody).toHaveBeenCalledWith("confirm-register", { firstname: data.firstname, lastname: data.lastname, activateLink: "http://localhost/#/authentication/activate?token=token" });
-    expect(sendMail).toHaveBeenCalledWith({ to: data.email, subject: "Confirmation d'inscription", html: "document" });
-  });
-});
