+++ api/tests/shared/acceptance/user-route.spec.js
@@ -3,7 +3,7 @@ import { describe, expect, it } from "vitest";
 
 import { databaseBuilder } from "../../../db/database-builder/index.js";
 import server from "../../../server.js";
-import { encodedToken } from "../../../src/authentication/services/token.service.js";
+import { encodedToken } from "../../../src/authentication/services/token-service.js";
 
 describe("Acceptance | Shared | User", () => {
   describe("GET /api/authenticated", () => {
